ISSUE Reporte comité técnico Contrucción del reporte de comité técnico

+++ backend/nsf/Code/Java/com/nutresa/exemplary_provider/dtl/queries/ReportOfAverageGradeBySuppliers.java
@@ -309,7 +309,7 @@ public class ReportOfAverageGradeBySuppliers {
         @Expose
         public String comment;
         @Expose
-        public short total;
+        public double total;
         @Expose
         public List<Item> items;
     }
